v1/functions/post_roles/requirements.txt
orator
PyMySQL
PyJWT
requests
